+++ udata_front/theme/gouvfr/assets/js/composables/admin/useCurrentOrganization.ts
@@ -1,0 +1,15 @@
+import type { Organization } from "@datagouv/components/ts";
+import { readonly, ref } from "vue";
+
+const currentOrganization = ref<Organization | null>(null);
+
+export function useCurrentOrganization() {
+  const setCurrentOrganization = (organization: Organization) => {
+    currentOrganization.value = organization;
+  }
+
+  return {
+    currentOrganization: readonly(currentOrganization),
+    setCurrentOrganization,
+  };
+}