+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Add index page with search for dataservices [#516](https://github.com/datagouv/udata-front/pull/516)
+- Add beta admin datasets page [#371](https://github.com/etalab/udata-front/pull/371)
 
 ## 5.2.0 (2024-09-13)